chore: move helper-functions.ts to /utils
